+++ Frameworks/Core/ERExtensions/Sources/er/extensions/formatters/ERXDividingNumberFormatter.java
@@ -92,7 +92,7 @@ public class ERXDividingNumberFormatter extends NSNumberFormatter {
             return super.format(b, arg1, arg2);
         } else {
             double d = n.doubleValue();
-            d /= (double)factor;
+            d /= factor;
             return super.format(Double.valueOf(d), arg1, arg2);
         }
     }
